[CLN] google_calendar: split test file
