+++ addons/google_calendar/tests/test_google_event.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from odoo.tests.common import BaseCase
 from odoo.addons.google_calendar.utils.google_calendar import GoogleEvent
