+++ addons/google_calendar/tests/test_token_access.py
@@ -1,3 +1,5 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from odoo import fields, Command
 from odoo.exceptions import AccessError
 from odoo.tests.common import TransactionCase
